requirements.txt
Django==1.6.2
South==0.8.4
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-jfu==2.0.1
django-summernote==0.5.6
django-toolbelt==0.0.1
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
python-dateutil==2.2
six==1.6.1
static==1.0.2
wsgiref==0.1.2
